Files not correctly added/removed from index on Windows platform
Updated concatenation of filepath and filename to prevent double backslash

On windows machines the last modified date is not updated when you copy a file.
For copied files the creation date has to be used. This does not cover renaming
files. Using java.nio library to get file creation date. This requires java 7

Closes #32.
Closes #60.
Closes #65.