Security: Update to Tika 1.5 / POI 3.10.1
There was an Apache POI release to address a security vulnerability.  For some document types, the fsriver plugin will indirectly use POI. This plugin release forces an update to Apache POI and is a response to the POI issue. Previously, the fsriver did not have an explicit dependency on POI. With this release, we have added a direct dependency and set it to the recent versions of POI. This will help users of the fsriver who might be unaware of these vulnerabilities, avoid them.

You can read more about the reported issues in [CVE-2014-3529](http://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2014-3529) and [CVE-2014-3574](http://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2014-3574)

We encourage anyone using the fsriver plugin with untrusted documents to update the plugin.

Closes #85.