Update README.md
With a fresh single node install of ES 2.2, the json exemple is OK if we replace
"port" : 9300
by
"port" : 9200.
As port 9200 is the default http port and port 9300 is the transport port, is that means that fscrawler doesn't use transport? (I'm a newbie to ES).
If so  lines from 874 to 889 should be corrected too.