Split tests into unit tests and real integration tests
To be able to test the same code against multiple versions of elasticsearch, we need to externalize our integration tests.

Which means:

* Download the right artifact we want to test, like `elasticsearch-2.3.2.zip`
* In pre-integration phase:
   * Unzip it
   * Launch `bin/elasticsearch`
* In integration phase, launch all `*IT` tests against this cluster
* In post-integration phase, stop the external cluster

To accomplish that, we need to:

* Rename integration tests from `*ITest` to `*IT` so it conforms to Maven guidelines
* Remove usage of elasticsearch `TransportClient` in our tests as it's a binary protocol
* Keep the possibility to run tests from the IDE (for debugging purposes). So if anyone wants to run tests from IDE, he will have to launch manually a cluster running at 9400 REST port. Like with `bin/elasticsearch -Des.http.port=9400`

Note that we removed tests against a 2 nodes cluster. This test is useful if the goal is to test the REST Client failover and load balancer mechanism but as the goal is to move to the official REST client, there won't be barely no need anymore to test an external library when #172 will be in.

This fix also fixes all the stuff added in #165.

Closes #175.
Closes #165.