Define a default exclusion list with files starting with `~`
Until now we were hard coding what is needed to excluded as known temporary files.
This commit makes that more flexible by defining a default exclusion list which can be overriden and
only defaults to files starting with `~` instead of files containing `~`.

So `excludes` is now set by default for new jobs to `["~*"]`. In previous versions, any file or directory containing a
`~` was excluded. Which means that if in your jobs, you are defining any exclusion rule, you need to add `*~*` if
you want to get back the exact previous behavior.

Closes #291.