Support `filename_as_id` for any type of file
Instead of supporting `filename_as_id` only for xml and json, we now support it for whatever file.
Note that it can cause some issues to the user as elasticsearch can't accept any kind of `_id`.

But it's on user's responsability to activate this option.

We can think of replacing on the fly any non valid character with a `_` for example but let's do that in another PR.

Closes #282.