Add new `add_as_inner_object` option to add metadata to json/xml documents
The default settings store the contents of json and xml documents directly onto the _source element of elasticsearch
documents. Thereby, there is no metadata about file and path settings, which are necessary to determine if a document
is deleted or updated.
New files will however be added to the index, (determined by the file timestamp).

If you need to keep json or xml documents synchronized to elasticsearch, you should set this option.

```json
{
  "name" : "test",
  "fs" : {
    "add_as_inner_object" : true
  }
}
```

Closes #237.
Closes #241.