Thread Leak in rest tests
https://travis-ci.org/dadoonet/fscrawler/builds/210289704

Basically the method just starts a REST interface and closes it immediately. I think that if we run some tests, that would be more useful and that would reduce the possibility of leaked threads.

On the other end I feel like it's something I should probably report in Glassfish project.

Closes #343.