Avoid bulk queue rejection
This commit adds a Retry mechanism when it detects that something wrong
happened with a document while indexing it.
It will slow down the injection process up to `30s`.

Note that it's not super easy to test it. I ran manual tests by starting
an elasticsearch instance with the following `elasticsearch.yml` settings:

```yml
thread_pool.bulk.size: 1
thread_pool.bulk.queue_size: 2
```

Closes #339