Reduce Time to run tests from the IDE
With recent changes like #500 and #514 it takes much more time to run all tests from the IDE when elasticsearch is not running yet locally.
The reason is that we start a new elasticsearch test container for every class. Starting a container can take some time.

We should try to reduce that and share the same docker instance for all tests.

Closes #516