Add tests on OSS image as well
Previously we were testing the code with x-pack and without.
We can still do that by pulling the right Docker image anytime we want to start integration tests.

This is available with a new Maven profile `oss`.

We also move the `docker-maven-plugin` to the parent pom so we can reuse this configuration in other modules.
There is not other one yet but I'm thinking of building at the end a FSCrawler Docker image which I can start for real and run more advanced integration tests.