Close bulk processors with awaitClose instead of close
We were closing the bulk processors with `close()` which was
trying to immediately exit the bulk processor.

Calling `awaitClose(30, TimeUnit.SECONDS)` is giving 30 more seconds
to the bulk processor to flush first all existing request before
actually closing.

I thought that `close()` was doing the same behind the scene but
apparently not.

Closes #547.