Make default root dir Windows compatible
By default, we are generating a `url` for the configuration as `/tmp/es`.
When we run the first time the crawler, FSCrawler tries to index
this root dir in elasticsearch as the root dir.

When doing so in `FsParser.indexDirectory()`, we are comparing `/tmp/es` with `File.separator`.
On Windows, the `File.separator` is `\` which leads to an error like:

```
22:34:17,188 WARN  [f.p.e.c.f.FsParser] Full stacktrace
java.lang.StringIndexOutOfBoundsException: String index out of range: -1
	at java.lang.String.substring(Unknown Source) ~[?:1.8.0_181]
	at fr.pilato.elasticsearch.crawler.fs.FsParser.indexDirectory(FsParser.java:550) ~[fscrawler-core-2.5.jar:?]
	at fr.pilato.elasticsearch.crawler.fs.FsParser.run(FsParser.java:160) [fscrawler-core-2.5.jar:?]
	at java.lang.Thread.run(Unknown Source) [?:1.8.0_181]
```

This commit changes the default generated configuration to either `/tmp/es` on non Windows systems or to `c:\tmp\es` on Windows systems.

Closes #533.