Exit as soon as we close FSCrawler
As reported in #605, we can not stop FSCrawler while it is crawling a directory.
There is no easy way to test this behavior automatically but manually.

This was caused by the fact that when we were closing FSCrawler we were then waiting for the wait time.
By default, we were waiting then for 10 minutes before FSCrawler can actually exit.

With this commit we now control before waiting for the wait time if FSCrawler is actually closing.
In which case we don't make the Crawler Thread wait.

This has been tested manually and seems to be working well.

Closes #605.