Change log level and display logs only if filters on content
This has been reported by a user on discuss.elastic.co: https://discuss.elastic.co/t/fscrawler-qui-crash-elasticsearch/172759/9

We can see in the logs that have been shared privately that the OOM is produced by log4j when we try to log:

```java
logger.debug("content = [{}], filters = {}", content, filters);
```

This was done in the very first line of `isIndexable(String content, List<String> filters)` and with a `DEBUG` level.
That could lead to OOM so I'm moving it to `TRACE` which is less used and this is now called only if we do have filters.