Allow passing test parameters to Maven CLI
What we documented about our tests using options like `tests.cluster.url`, `tests.cluster.cloud_id`, `tests.cluster.user` and `tests.cluster.pass` was actually working only from the IDE but not when using Maven CLI.
This change makes sure we are passing all those env variables to the randomizedtesting maven plugin.

We are also doing that with `tests.rest.port` which is yet undocumented but will be soon with #728.

Closes #727.