Fix flaky tests
Sometimes the shards do not seem to be ready for querying just after the index creation and we are failing the suite.
It sounds like Elasticsearch can throw an `ElasticsearchException` which is a `RuntimeException` and not an `IOException`.

This change tries to catch it so we can retry before failing.