Switch to Docker-Compose instead of Docker for tests
We are switching from `docker-maven-plugin` to `docker-compose-maven-plugin` to have a better way to test our full platform.
That way we can also provide to the community a docker-compose contrib example which we are using within our own tests.

Note that this is using a not released yet version for the maven plugin waitfor: https://github.com/simplaex/waitfor-maven-plugin/