Fix flaky tests
When we are returning a negative value in the tests, we considered this as the final result instead of waiting for another value.

That was wrong.

This commit fixes it and adds a test to make sure we don't have a regression in the future.

This should definitely closes issues like #679.

Was the last part missing in #971.