Disable automatic integration tests with Workplace Search
We don't have an Admin API and all the tests to fake this API are pretty much unstable.

As I want to release a version soonish, we will be running the IT with Workplace Search only manually using the CLI options `-Dtests.workplace.access_token=XXX -Dtests.workplace.key=YYY`.

We might test all that using a cloud always running instance...