Detect when fscrawler runs in foreground and adapt logs
To ease the debugging, we now detect that fscrawler is launched in foreground or background.

When in background, we use a specific file only logger (`log4j2-file.xml`).

When in foreground, we use the standard logger  (`log4j2.xml`) which prints both to the console and in the logs dir.