Do not hide YAML parsing errors
When parsing a wrong yaml file, we are today catching the error and behave as if the file was not existing and proposes the user to create the missing file:

```
job [jobname] does not exist
Do you want to create it (Y/N)?
```

This is leading users to wrong assumptions.

This change now prints a meaningful error while exiting:

```
13:05:24,021 FATAL [f.p.e.c.f.c.FsCrawlerCli] Cannot parse the configuration file: while parsing a block mapping
 in 'reader', line 1, column 1:
    name: "test"
    ^
expected <block end>, but found '<block mapping start>'
 in 'reader', line 4, column 2:
     follow_symlink: false
     ^

 at [Source: (StringReader); line: 4, column: 2]
```

Closes #1767.