Remove workplace search integration
Leftovers. Related to #1823 and #1820.

See the text [in the documentation](https://www.elastic.co/guide/en/workplace-search/8.12/index.html)

> For new users, we recommend using our Elasticsearch-native tools, rather than the standalone Workplace Search product. See [this blog post](https://www.elastic.co/blog/evolution-workplace-search-private-data-elasticsearch) for more information about upgrading your internal knowledge search, to make it an amazing experience for your users!

So this means that we should just be sending documents as we did in the past to Elasticsearch, directly.
There's a connector framework though but it's only available in Python and is not usable here.

Closes #1820.