Add support for password protected documents
This also change the way the Tika parser was instantiated. It's no more a static class.

This is a WIP as I'd like to add the support for multiple passwords so we can try many options "a la brute force" in case the directory contains many files with different passwords.

I think we should implement a PasswordProvider interface which could get the Password from many possible providers.

The idea is to define `PasswordProvider#getPassword(String path)` which is responsible to provide the password for a given file.

The simplest one would be `MemoryPasswordProvider`.
The easiest one would be `DiskPasswordProvider`.
And may be an `ElasticsearchPasswordProvider`.

Related to #1916.